Replaced "which see" with "q.v.".
While it may seem strange to replace an English phrase with a
relatively obscure Latin abbreviation, there are a number of
advantages.  As someone whose first language is English, I've never
seen "which see" used anywhere else the way it is in Emacs.  It looks
totally ungrammatical to me, and I found it quite confusing when I
first encountered it.  It's a difficult phrase to search for because
it's formed from such common words.  OTOH, q.v. is easily searchable,
it has a well-defined meaning, and the conventional way to use it
serves as a drop-in replacement for the way "which see" is used in
Emacs.